+++ app/src/main/java/com/example/licentaincercarea1/MainActivity.kt
@@ -1,6 +1,7 @@
 package com.example.licentaincercarea1
 
 
+import android.content.Intent
 import android.content.res.Resources
 import android.os.Bundle
 import android.util.Log
@@ -17,7 +18,7 @@ import com.google.android.material.navigation.NavigationView
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
-    private lateinit var appBarConfiguration : AppBarConfiguration
+
     private lateinit var binding: ActivityMainBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -26,56 +27,14 @@ class MainActivity : AppCompatActivity() {
         setContentView(binding.root)
         setListener()
 
-
-
-        val host: NavHostFragment = supportFragmentManager
-            .findFragmentById(R.id.my_nav_host_fragment) as NavHostFragment? ?: return
-
-        val navController = host.navController
-
-
-
-        setupNavigationMenu(navController)
-
-        setupBottomNavMenu(navController)
-
-        navController.addOnDestinationChangedListener { _, destination, _ ->
-            val dest: String = try {
-                resources.getResourceName(destination.id)
-            } catch (e: Resources.NotFoundException) {
-                Integer.toString(destination.id)
-            }
-
-            Log.d("NavigationActivity", "Navigated to $dest")
-        }
-
     }
     private fun setListener() {
 
         binding.button.setOnClickListener {
-            supportFragmentManager.beginTransaction().replace(R.id.container, CategoriesFragment()).commit()
-            button.visibility= View.GONE
+            startActivity(Intent(this@MainActivity,FragmentBase::class.java))
         }
 
     }
 
-    private fun setupBottomNavMenu(navController: NavController) {
-        val bottomNav = findViewById<BottomNavigationView>(R.id.bottom_nav_view)
-        bottomNav?.setupWithNavController(navController)
-    }
-
-    private fun setupNavigationMenu(navController: NavController) {
-        val sideNavView = findViewById<NavigationView>(R.id.nav_view)
-        sideNavView?.setupWithNavController(navController)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return item.onNavDestinationSelected(findNavController(R.id.my_nav_host_fragment))
-                || super.onOptionsItemSelected(item)
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        return findNavController(R.id.my_nav_host_fragment).navigateUp(appBarConfiguration)
-    }
 
 }
