+++ app/src/main/res/layout-w960dp/navigation_activity.xml
@@ -24,21 +24,12 @@
         android:layout_toEndOf="@id/nav_view"
         android:background="?android:attr/listDivider" />
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_toEndOf="@id/nav_view"
-        android:background="@color/white"
-        android:theme="@style/ThemeOverlay.MaterialComponents.Dark.ActionBar" />
 
     <fragment
         android:id="@+id/my_nav_host_fragment"
         android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@id/toolbar"
         android:layout_toEndOf="@id/nav_view"
         app:defaultNavHost="true"
         app:navGraph="@navigation/app_nav" />
