+++ app/src/main/res/menu/nav_drawer_menu.xml
@@ -4,11 +4,17 @@
     <group android:id="@+id/primary">
         <item
             android:id="@id/categoriesFragment"
-            android:icon="@drawable/food_icon"
-            android:title="@string/item_title"/>
+            android:icon="@drawable/categorii"
+            android:title="@string/categorii"/>
         <item
             android:id="@id/manifesting"
-            android:icon="@drawable/food_icon"
-            android:title="@string/item_des" />
+            android:icon="@drawable/retete"
+            android:title="@string/retete"
+            />
+        <item
+            android:id="@id/manifesting2"
+            android:icon="@drawable/fav"
+            android:title="@string/favorie"
+            />
     </group>
 </menu>
